requirements.txt
ollama==0.4.7
faiss-cpu==1.9.0
sentence-transformers==3.4.0
pyaudio==0.2.14
numpy==1.26.4
whisper==1.1.10
TTS==0.22.0
sounddevice==0.5.1

